requirements.txt
Django
djangorestframework
mysqlclient
celery[redis]
django-redis